Make navigation stateful
This commit introduces :selected in the 'search' Vue route, so we can
store which node is selected within a search result. Both the nodes and
the filenames (cer/mft/crl/roa) in the visualisation have full links so
they can be opened in separate tabs.
Furthermore a 'Selected' tab is introduced in the Detail view, to
distinguish selected ROAs from other nodes clicked by the user.
Page titles now also include the search query and the basename of the
selected node, improving navigating back/forward.

Implementation-wise, the URL is leading when it comes to the state.